Initial settings structure
Changed the initial settings structure so that iStarter.settings looks
for a config file which tells it which environment it's running in.

I've built a settings file for each of us to modify according to our
own development environments and another couple of deployment envs.

Note, that I'm also using a settings file in the ideasapp so that some
settings stay with the app (for easier portability). The app-level
settings file inherits from the project-level settings file.